refactor: rename ResultSearchCard to PartCard
