+++ src/components/PartCard/ResultSearchCardSkeleton.tsx
