+++ src/components/PartCard/SearchInfo/styles.ts
