+++ src/components/PartCard/SearchInfoCompact.tsx
