+++ src/services/hooks/list/useGetListById.tsx
@@ -1,6 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
 import apiService from '../../api';
 
+/**
+ *
+ * @param listId
+ * @returns
+ */
 export const useGetListById = (listId: string) => {
     return useQuery({
         queryKey: ['list', listId],
